feat: Loading Parquet data via `workerpool`
