+++ package.json
@@ -60,6 +60,7 @@
     "react-router-dom": "^6.22.2",
     "styled-components": "^6.0.8",
     "three": "^0.151.3",
-    "usehooks-ts": "^2.9.1"
+    "usehooks-ts": "^2.9.1",
+    "workerpool": "^9.1.3"
   }
 }
